requirements.txt
django==4.2.11
psycopg2-binary==2.9.9
pillow==10.2.0
django-storages==1.14.2
boto3==1.34.41
isort==5.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1